Mostly Graphing updates and database schema change

+++ src/com/bryanmarty/greenbutton/data/DataPoint.java
@@ -3,9 +3,9 @@ package com.bryanmarty.greenbutton.data;
 public class DataPoint {
 
 	public int x;
-	public int y;
+	public double y;
 	
-	public int value;
-	public int cost;
+	public double value;
+	public double cost;
 	
 }
